+++ steel/src/steel_vm/heap.rs
@@ -1,435 +1,146 @@
-use crate::{env::Env, gc::OBJECT_COUNT, rvals::SteelVal};
+use crate::{rvals::UpValue, SteelVal};
 use std::cell::RefCell;
-use std::rc::Rc;
-use std::rc::Weak;
-
 use std::collections::HashMap;
-pub(crate) static HEAP_LIMIT: usize = 100;
+use std::rc::{Rc, Weak};
 
-use log::debug;
+const GC_THRESHOLD: usize = 100;
+const GC_GROW_FACTOR: usize = 2;
+const _RESET_LIMIT: usize = 5;
 
-/// Garbage collection central
-/// The purpose of this is simply to hold environments that get allocated
-pub struct Heap {
-    heap: Vec<Rc<RefCell<Env>>>,
-    root: Option<Weak<RefCell<Env>>>,
-    limit: usize,
-    max_double: u8,
-    current_double: u8,
-    // Use for deciding when to run a full traversal
-    purge_count: u8,
-    purge_max: u8,
+pub struct UpValueHeap {
+    memory: Vec<Rc<RefCell<UpValue>>>,
+    count: usize,
+    threshold: usize,
 }
 
-impl Default for Heap {
-    fn default() -> Self {
-        Heap {
-            heap: Vec::new(),
-            root: None,
-            limit: HEAP_LIMIT,
-            max_double: 2,
-            current_double: 0,
-            purge_count: 0,
-            purge_max: 5,
-        }
-    }
-}
-
-impl Drop for Heap {
-    fn drop(&mut self) {
-        self.heap.clear();
-    }
-}
-
-impl Heap {
+impl UpValueHeap {
     pub fn new() -> Self {
-        Heap {
-            heap: Vec::new(),
-            root: None,
-            limit: HEAP_LIMIT,
-            max_double: 2,
-            current_double: 0,
-            purge_count: 0,
-            purge_max: 5,
+        UpValueHeap {
+            memory: Vec::new(),
+            count: 1,
+            threshold: GC_THRESHOLD,
         }
     }
 
-    pub fn plant_root(&mut self, root_ref: Weak<RefCell<Env>>) {
-        self.root.replace(root_ref);
-    }
-
-    // pub fn add(&mut self, val: Rc<RefCell<Env>>) {
-    //     if !val.borrow().is_root() {
-    //         self.heap.push(val)
-    //     }
-    // }
-
-    pub fn clear(&mut self) {
-        debug!("Freeing the heap with length: {}", self.len());
-        self.heap.clear()
-    }
-
-    pub fn limit(&self) -> usize {
-        self.limit
-    }
-
-    // pub fn _inspect_heap(&self) {
-    //     println!("heap length: {}", self.heap.len());
-    //     self.heap.iter().for_each(|x| {
-    //         println!(
-    //             "{}, {}, {}",
-    //             x.borrow().is_reachable(),
-    //             Rc::weak_count(x),
-    //             Rc::strong_count(x)
-    //         )
-    //     });
-    //     let hp: Vec<String> = self
-    //         .heap
-    //         .iter()
-    //         .map(|x| x.borrow().string_bindings_vec())
-    //         .collect();
-    //     println!("{:?}", hp);
-    // }
-
-    pub fn len(&self) -> usize {
-        self.heap.len()
-    }
-
-    // pub fn is_empty(&self) -> bool {
-    //     self.heap.is_empty()
-    // }
-
-    // pub fn reset_limit(&mut self) {
-    //     self.limit = 5000;
-    // }
-
-    pub fn append(&mut self, other: &mut Self) {
-        self.heap.append(&mut other.heap)
-    }
-
-    // pub fn collect_garbage(&mut self) {
-    //     // println!("Calling garbage collection");
-
-    //     if self.len() > self.limit {
-    //         // std::thread::sleep(std::time::Duration::new(3, 0));
-    //         debug!(
-    //             "Before mark and sweep - Heap-length: {}, Active-Object-Count: {:?}",
-    //             self.len(),
-    //             OBJECT_COUNT
-    //         );
-    //         // self.profile_heap();
-    //         self.drop_large_refs();
-    //         if self.current_double < self.max_double {
-    //             self.limit *= 2;
-    //             self.current_double += 1;
-    //         } else {
-    //             // std::thread::sleep(std::time::Duration::new(3, 0));
-    //             // println!("******************************************************");
-    //             // println!("******************* RESET ****************************");
-    //             // println!("******************************************************");
-    //             self.limit = HEAP_LIMIT;
-    //             self.current_double = 0;
-    //             // std::thread::sleep(std::time::Duration::new(3, 0));
-    //         }
-    //         // self.profile_heap();
-
-    //         self.purge_count += 1;
-
-    //         debug!(
-    //             "After mark and sweep - Heap-length: {}, Active-Object-Count: {:?}",
-    //             self.len(),
-    //             OBJECT_COUNT
-    //         );
-
-    //         debug!("Heap limit set to: {}", self.limit);
-
-    //         if self.purge_count == self.purge_max {
-    //             println!("big purge");
-    //             if let Some(root) = &self.root {
-    //                 let root = root.upgrade().unwrap();
-    //                 self.gather_big_mark_and_sweep(&root);
-    //             }
-    //             self.purge_count = 0;
-    //         }
-
-    //         // std::thread::sleep(std::time::Duration::new(3, 0));
-    //     }
-    // }
-
-    pub fn profile_heap(&self) {
-        let mut hm = HashMap::new();
-        for val in &self.heap {
-            let key = (Rc::weak_count(val), Rc::strong_count(val));
-            // hm.insert()
-            let count = hm.entry(key).or_insert(0);
+    fn _profile_heap(&self) {
+        let mapped = self
+            .memory
+            .iter()
+            .map(|x| Rc::weak_count(x))
+            .collect::<Vec<_>>();
+        let mut hm: HashMap<usize, usize> = HashMap::new();
+        for value in mapped {
+            // hm.insert(k, v)
+            let count = hm.entry(value).or_insert(0);
             *count += 1;
         }
 
-        debug!("Heap profile: {:?}", hm);
+        println!("{:#?}", hm);
     }
 
-    pub fn drop_large_refs(&mut self) {
-        debug!("Dropping envs with a weak count of 0");
+    // This does not handle cycles, perhaps add an explicit cycle detection via traversal
+    // to mark things reachable?
+    fn collect<'a>(&mut self, roots: impl Iterator<Item = &'a SteelVal>) {
+        if self.memory.len() > self.threshold {
+            // let prior = self.memory.len();
 
-        // self.heap
-        //     .retain(|x| Rc::weak_count(x) > 1 && Rc::strong_count(x) > 1);
+            let mut changed = true;
+            while changed {
+                let prior_len = self.memory.len();
+                self.memory.retain(|x| Rc::weak_count(x) > 0);
+                let after = self.memory.len();
+                changed = prior_len != after;
+            }
+            // self.memory.retain(|x| Rc::weak_count(x) > 0);
 
-        // self.heap
-        //     .retain(|x| (Rc::weak_count(x) + x.borrow().weak_count()) > 1);
-        self.heap.retain(|x| Rc::weak_count(x) > 0);
-        // Drop the heap size back down to conserve memory
-        self.heap.shrink_to_fit();
+            self.mark_and_sweep(roots);
 
-        // self.heap.retain(|x| Rc::strong_count(x) > 1);
+            // let after = self.memory.len();
+
+            // TODO check this
+            self.threshold = (self.threshold + self.memory.len()) * GC_GROW_FACTOR;
+
+            // println!("new threshold: {}", self.threshold);
+
+            // self.profile_heap();
+            // println!(
+            //     "Freed: {}, New heap length: {}",
+            //     prior - after,
+            //     self.memory.len()
+            // );
+            self.count += 1;
+        }
     }
 
-    // #[inline]
-    // fn match_closure(&mut self, val: &SteelVal) {
-    //     match val {
-    //         SteelVal::Closure(bytecode_lambda) => {
-    //             let p_env = bytecode_lambda.sub_expression_env().upgrade().unwrap();
+    fn mark_and_sweep<'a>(&mut self, roots: impl Iterator<Item = &'a SteelVal>) {
+        // mark
+        for root in roots {
+            traverse(root);
+        }
 
-    //             self.big_gather(&p_env);
+        // sweep
+        self.memory
+            .retain(|x| x.borrow().is_reachable() || !x.borrow().is_closed());
 
-    //             let reachable = p_env.borrow().is_reachable();
+        // put them back as unreachable
+        self.memory.iter().for_each(|x| x.borrow_mut().reset());
+    }
 
-    //             if !reachable {
-    //                 self.big_gather(&p_env);
-    //             }
-    //         }
-    //         SteelVal::Pair(_) => {
-    //             // println!("Getting here!");
-    //             SteelVal::iter(val.clone()).for_each(|x| self.match_closure(&x))
-    //         }
-    //         SteelVal::VectorV(v) => v.iter().for_each(|x| self.match_closure(x)),
-    //         SteelVal::HashMapV(hm) => hm.values().for_each(|x| self.match_closure(x)),
-    //         SteelVal::HashSetV(hs) => hs.iter().for_each(|x| self.match_closure(x)),
-    //         _ => {} // SteelVal::Closure
-    //     }
-    // }
+    pub(crate) fn new_upvalue<'a>(
+        &mut self,
+        index: usize,
+        next: Option<Weak<RefCell<UpValue>>>,
+        roots: impl Iterator<Item = &'a SteelVal>,
+    ) -> Weak<RefCell<UpValue>> {
+        let upvalue = Rc::new(RefCell::new(UpValue::new(index, next)));
+        let weak_ptr = Rc::downgrade(&upvalue);
+        self.memory.push(upvalue);
 
-    // pub fn big_gather(&mut self, leaf: &Rc<RefCell<Env>>) {
-    //     self.add(Rc::clone(leaf));
+        // self.profile_heap();
+        self.collect(roots);
+        // self.profile_heap();
 
-    //     if leaf.borrow().is_root() {
-    //         return;
-    //     }
+        weak_ptr
+    }
+}
 
-    //     let mut env = Rc::clone(leaf);
-    //     // let mut heap = vec![Rc::clone(leaf)];
+// Use this function to traverse and find all reachable things
+// 'reachable' should be values living in the heap, stack, and in the
+fn traverse(val: &SteelVal) {
+    match val {
+        SteelVal::Pair(_) => {}
+        SteelVal::VectorV(_) => {}
+        SteelVal::HashMapV(_) => {}
+        SteelVal::HashSetV(_) => {}
+        SteelVal::StructV(_) => {}
+        SteelVal::PortV(_) => {}
+        SteelVal::Closure(c) => {
+            for upvalue in c.upvalues() {
+                let upvalue = upvalue.upgrade().unwrap();
+                mark_upvalue(&upvalue);
+            }
+        }
+        SteelVal::IterV(_) => {}
+        SteelVal::FutureV(_) => {}
+        SteelVal::StreamV(_) => {}
+        SteelVal::BoxV(_) => {}
+        SteelVal::Contract(_) => {}
+        SteelVal::ContractedFunction(_) => {}
+        SteelVal::ContinuationFunction(_) => {}
+        _ => {
+            // println!("Doesn't need to be traversed")
+        }
+    }
+}
 
-    //     while let Some(parent_env) = Rc::clone(&env).borrow().sub_expression() {
-    //         let upgraded_env = parent_env.upgrade().unwrap();
-    //         if !upgraded_env.borrow().is_root() {
-    //             let e = Rc::clone(&upgraded_env);
-    //             self.add(Rc::clone(&e));
-    //             // See if this works...
-    //             // TODO come back to this
-    //             // MATCH CLOSURE PROBLEM
-    //             for value in e.borrow().bindings_map().values() {
-    //                 self.match_closure(value);
-    //             }
+#[inline(always)]
+fn mark_upvalue(upvalue: &Rc<RefCell<UpValue>>) {
+    // unimplemented!()
+    {
+        upvalue.borrow_mut().mark_reachable();
+    }
 
-    //             env = upgraded_env;
-    //         } else {
-    //             break;
-    //         }
-    //     }
-    // }
-
-    // pub fn gather(&mut self, leaf: &Rc<RefCell<Env>>) {
-    //     self.add(Rc::clone(leaf));
-
-    //     if leaf.borrow().is_root() {
-    //         return;
-    //     }
-
-    //     let mut env = Rc::clone(leaf);
-    //     // let mut heap = vec![Rc::clone(leaf)];
-
-    //     while let Some(parent_env) = Rc::clone(&env).borrow().sub_expression() {
-    //         let upgraded_env = parent_env.upgrade().unwrap();
-    //         if !upgraded_env.borrow().is_root() {
-    //             let e = Rc::clone(&upgraded_env);
-    //             self.add(Rc::clone(&e));
-    //             // See if this works...
-    //             // TODO come back to this
-    //             // MATCH CLOSURE PROBLEM
-    //             // for value in e.borrow().bindings_map().values() {
-    //             //     self.match_closure(value);
-    //             // }
-
-    //             env = upgraded_env;
-    //         } else {
-    //             break;
-    //         }
-    //     }
-    // }
-
-    // pub fn mark(&self) {
-    //     self.heap
-    //         .iter()
-    //         .for_each(|x| x.borrow_mut().set_reachable(true));
-    // }
-
-    // pub fn reset(&self) {
-    //     self.heap
-    //         .iter()
-    //         .for_each(|x| x.borrow_mut().set_reachable(false));
-    // }
-
-    // pub fn drop_unreachable(&mut self) {
-    //     &self.heap.retain(|x| x.borrow().is_reachable());
-    // }
-
-    // pub fn sweep(&mut self) {
-    //     // std::thread::sleep(std::time::Duration::new(5, 0));
-    //     // println!(
-    //     //     "env currently at sweep: {:?}",
-    //     //     self.heap
-    //     //         .iter()
-    //     //         .map(|x| (
-    //     //             x.borrow().is_reachable(),
-    //     //             Rc::weak_count(x),
-    //     //             x.borrow().weak_count()
-    //     //         ))
-    //     //         .collect::<Vec<_>>()
-    //     // );
-    //     // std::thread::sleep(std::time::Duration::new(5, 0));
-
-    //     &self
-    //         .heap
-    //         .retain(|x| x.borrow().is_reachable() || Rc::weak_count(x) > 0);
-    //     // .retain(|x| x.borrow().is_reachable());
-    // }
-
-    // pub fn mark_and_sweep(&mut self) {
-    //     self.mark();
-    //     self.sweep();
-    // }
-
-    // pub fn collect_mark_and_sweep(&mut self, leaf: &Rc<RefCell<Env>>) {
-    //     self.gather(leaf);
-    //     self.mark_and_sweep();
-    // }
-
-    // pub fn gather_and_mark(leaf: &Rc<RefCell<Env>>) {
-    //     let mut heap = Self::new();
-    //     heap.gather(leaf);
-    //     heap.mark();
-    // }
-
-    // pub fn gather_and_mark_2(leaf1: &Rc<RefCell<Env>>, leaf2: &Rc<RefCell<Env>>) {
-    //     let mut heap = Self::new();
-    //     heap.gather(leaf1);
-    //     heap.gather(leaf2);
-    //     heap.mark();
-    // }
-
-    // pub fn gather_mark_and_sweep(&mut self, leaf: &Rc<RefCell<Env>>) {
-    //     Self::gather_and_mark(leaf);
-    //     self.sweep();
-    //     self.reset();
-    // }
-
-    // pub fn gather_big_mark_and_sweep(&mut self, root: &Rc<RefCell<Env>>) {
-    //     debug!("Running full mark and sweep");
-
-    //     {
-    //         let mut heap = Self::new();
-    //         for value in root.borrow().bindings_map().values() {
-    //             heap.match_closure(value);
-    //         }
-    //         heap.mark();
-    //         // dbg!(&heap.heap);
-    //     }
-    //     // dbg!(&self.heap);
-    //     self.drop_unreachable();
-    //     self.reset();
-    // }
-
-    // pub fn collect_garbage_with_leaves(
-    //     &mut self,
-    //     leaf1: &Rc<RefCell<Env>>,
-    //     leaf2: &Rc<RefCell<Env>>,
-    // ) {
-    //     // println!("Calling garbage collection");
-
-    //     if self.len() > self.limit {
-    //         self.gather_mark_and_sweep_2(leaf1, leaf2);
-
-    //         // std::thread::sleep(std::time::Duration::new(3, 0));
-    //         debug!(
-    //             "Before mark and sweep - Heap-length: {}, Active-Object-Count: {:?}",
-    //             self.len(),
-    //             OBJECT_COUNT
-    //         );
-    //         // self.profile_heap();
-    //         self.drop_large_refs();
-    //         if self.current_double < self.max_double {
-    //             self.limit *= 2;
-    //             self.current_double += 1;
-    //         } else {
-    //             // std::thread::sleep(std::time::Duration::new(3, 0));
-    //             // println!("******************************************************");
-    //             // println!("******************* RESET ****************************");
-    //             // println!("******************************************************");
-    //             self.limit = HEAP_LIMIT;
-    //             self.current_double = 0;
-    //             // std::thread::sleep(std::time::Duration::new(3, 0));
-    //         }
-    //         // self.profile_heap();
-
-    //         self.purge_count += 1;
-
-    //         debug!(
-    //             "After mark and sweep - Heap-length: {}, Active-Object-Count: {:?}",
-    //             self.len(),
-    //             OBJECT_COUNT
-    //         );
-
-    //         debug!("Heap limit set to: {}", self.limit);
-
-    //         if self.purge_count == self.purge_max {
-    //             // println!("big purge");
-
-    //             if let Some(root) = &self.root {
-    //                 let root = root.upgrade().unwrap();
-    //                 self.gather_big_mark_and_sweep(&root);
-    //             }
-    //             self.purge_count = 0;
-    //         }
-
-    //         // std::thread::sleep(std::time::Duration::new(3, 0));
-    //     }
-    // }
-
-    // pub fn gather_mark_and_sweep_2(&mut self, leaf1: &Rc<RefCell<Env>>, leaf2: &Rc<RefCell<Env>>) {
-    //     // println!(
-    //     //     "!!!!!!!!!!! ############# gather, mark and sweep ############### !!!!!!!!!!!!!!!"
-    //     // );
-
-    //     debug!("Running mark and sweep");
-    //     Self::gather_and_mark_2(leaf1, leaf2);
-    //     // self.sweep();
-
-    //     // if self.len() > self.limit {
-    //     //     debug!("Running mark and sweep");
-    //     //     Self::gather_and_mark_2(leaf1, leaf2);
-    //     //     self.sweep();
-    //     //     self.reset();
-    //     // }
-
-    //     // self.add(Rc::clone(leaf1));
-    //     // self.add(Rc::clone(leaf2));
-    // }
-
-    // pub fn gather_from_slice(&mut self, args: &[Gc<SteelVal>]) {
-    //     for arg in args {
-    //         if let SteelVal::Closure(closure) = arg.as_ref() {
-    //             let parent_env = closure.sub_expression_env().upgrade().unwrap();
-    //             self.gather(&parent_env)
-    //         }
-    //     }
-    // }
+    if let Some(inner) = upvalue.borrow().get_value_if_closed() {
+        traverse(inner);
+    }
 }
