+++ steel/src/steel_vm/mod.rs
@@ -2,8 +2,7 @@ pub(crate) mod const_evaluation;
 mod contracts;
 pub mod engine;
 mod evaluation_progress;
-pub(crate) mod heap;
-mod heap2;
+mod heap;
 mod inline_iter;
 mod lazy_stream;
 mod primitives;
